Rat/requirements.txt
# RAT BRAIN Requirements
# Install on Raspberry Pi with: pip3 install -r requirements.txt

# GPIO Control (Raspberry Pi)
RPi.GPIO>=0.7.0
# or use: gpiozero>=2.0

# NeoPixel / WS2812 LED Control
adafruit-neopixel>=1.12.0
# or use rpi-ws281x from local Libs

# Optional: PyQt5 for UI testing on PC
# PyQt5>=5.15.0

# Python 3.7+ (included in Raspberry Pi OS)
# socket, threading, time, logging, queue (all built-in)
